split RgbdRendererTest fixture to a header file, so it can also be used to test godot
