+++ systems/sensors/BUILD.bazel
@@ -447,6 +447,12 @@ drake_cc_googletest(
     ],
 )
 
+drake_cc_library(
+    name = "rgbd_renderer_test",
+    hdrs = ["test/rgbd_renderer_test.h"],
+    testonly = 1,
+)
+
 drake_cc_googletest(
     name = "rgbd_renderer_vtk_test",
     data = [
@@ -455,6 +461,16 @@ drake_cc_googletest(
     # Mitigates driver-related issues when running under `bazel test`. For more
     # information, see #7004.
     local = 1,
+    tags = ["manual"],
+    deps = [
+        ":rgbd_renderer",
+        "//drake/common:find_resource",
+        "//drake/common/test_utilities:eigen_matrix_compare",
+        "//drake/math:geometric_transform",
+        ":rgbd_renderer_test"
+    ],
+)
+
     # Disable under LeakSanitizer and Valgrind Memcheck due to driver-related
     # leaks. For more information, see #7520.
     # TODO(jwnimmer-tri) Marked manual because it opens windows on the
